+++ src/parser/mod.rs
@@ -1,1 +1,2 @@
+pub mod covariate_model;
 pub mod model_parser;